making openllama compatible with cpu

+++ models/open_llama.py
@@ -8,7 +8,7 @@ class HuggingAdapter:
     def __init__(self, configuration):
         model_path = configuration.get("model_path", 'openlm-research/open_llama_3b')
         self.tokenizer = LlamaTokenizer.from_pretrained(model_path)
-        self.model = LlamaForCausalLM.from_pretrained(model_path, torch_dtype=torch.float16, device_map='auto')
+        self.model = LlamaForCausalLM.from_pretrained(model_path, device_map='auto')
         self.top_k = configuration.get("top_k", 5)
     
     def prepare_item_func(self, item: dl.Item):
